feat: add permissions for writing to the documentation workflow
The permissions for writing to the documentation workflow have been added to ensure that the necessary access is granted for writing to the contents. This is necessary to enable the workflow to write to the required locations during the documentation process.

The changes also include the addition of a new file, `.latexindent.yaml`, which contains the default settings for indentation and horizontal space in the LaTeX files. This file is added to ensure consistent formatting of LaTeX files.